Fixed issues, only issue remaining is with the Title field
