+++ Source/EVEClassLibrary/MarketListing.cs
