+++ Source/EVEMarketeerTest/EVEMarketeerTest.csproj
@@ -17,7 +17,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\EVEClassLibrary\EVEClassLibrary\EVEClassLibrary.csproj" />
+    <ProjectReference Include="..\EVEClassLibrary\EVEClassLibrary.csproj" />
   </ItemGroup>
 
   <ItemGroup>
