--- Source/EVEMarketeerTest/EVEMarketeerTest.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.9.34728.123
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EVEMarketeerTest", "EVEMarketeerTest.csproj", "{A9F7B533-996C-4683-86DD-CBB86E120072}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A9F7B533-996C-4683-86DD-CBB86E120072}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A9F7B533-996C-4683-86DD-CBB86E120072}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A9F7B533-996C-4683-86DD-CBB86E120072}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A9F7B533-996C-4683-86DD-CBB86E120072}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {F5DD4ABC-8B9C-46C4-87F9-F40C0A77978A}
-	EndGlobalSection
-EndGlobal
