--- Source/Eve Console Tester/Eve Console Tester.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.9.34728.123
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Eve Console Tester", "Eve Console Tester.csproj", "{EA494E1A-B787-40A0-A3E1-836259CAF1F5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{EA494E1A-B787-40A0-A3E1-836259CAF1F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EA494E1A-B787-40A0-A3E1-836259CAF1F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EA494E1A-B787-40A0-A3E1-836259CAF1F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EA494E1A-B787-40A0-A3E1-836259CAF1F5}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {BB95C833-68FF-4620-B963-1E36E01A7801}
-	EndGlobalSection
-EndGlobal
